PN-17351

+++ PAS_Database/dbo/Tables/Tables1/AssetInventory.sql
@@ -116,7 +116,7 @@
     [Level4]                            VARCHAR (200)   NULL,
     [ManufactureName]                   VARCHAR (100)   NULL,
     [LocationName]                      VARCHAR (100)   NULL,
-    [Qty]                               INT             NULL,
+    [Qty]                               DECIMAL (18, 6) NULL,
     [StklineNumber]                     VARCHAR (100)   NULL,
     [AvailStatus]                       VARCHAR (100)   NULL,
     [PartNumber]                        VARCHAR (100)   NULL,
@@ -179,6 +179,8 @@
     CONSTRAINT [FK_AssetInventory_UnitOfMeasure] FOREIGN KEY ([UnitOfMeasureId]) REFERENCES [dbo].[UnitOfMeasure] ([UnitOfMeasureId]),
     CONSTRAINT [Unique_AssetInventory] UNIQUE NONCLUSTERED ([InventoryNumber] ASC, [MasterCompanyId] ASC)
 );
+
+
